reformat according to pep; allow multiple doctor-locator searching; bump requires versions; test under python 3.13
