+++ README.md
@@ -7,15 +7,20 @@ How to use LuxmedSniper?
 --------------------
 First of all create virtualenv and install Python requirements from requirements.txt
 
-1) For each specialist create configuration file (yaml format) and save it for example as my_favourite_surgeon.yml:
+1) For each luxmed users create configuration file (yaml format) and save it for example as luxmed_username.yml:
 ```
 luxmed:
   email: EMAIL
   password: PASSWORD
-luxmedsniper: #                     mandatory  mandatory
-  doctor_locator_id: 5*4430*-1*-1 # (cityId, serviceVariantId, facilitiesIds, doctorsIds) -1 means any.
-                                  # You can get those ids by reading form data sent to https://portalpacjenta.luxmed.pl/PatientPortal/Reservations/Reservation/PartialSearch
-                                  # on https://portalpacjenta.luxmed.pl/PatientPortal/Reservations/Reservation/Search by chrome dev tools
+luxmedsniper:
+  doctor_locators:
+    - id: 1*7409*-1*-1 # (cityId, serviceVariantId, facilitiesIds, doctorsIds) -1 means any.
+      # You can get those ids by calling script with "--dump-ids" argument: python3 luxmed_sniper.py --dump-ids
+      name: Your unique search name
+      enabled: False  # temporary disable from searching
+    - id: 1*7681*-1*-1
+      name: Your unique search name 2
+      enabled: True
   lookup_time_days: 14 # How many days from now should script look at.
 pushover:
   user_key: # Your pushover.net user key
@@ -23,16 +28,16 @@ pushover:
   message_template: "New visit! {AppointmentDate} at {ClinicPublicName} - {DoctorName}"
   title: "New Lux Med visit available!" # Pushover message topic
 misc:
-  notifydb: ./surgeon_data # State file used to remember which notifications has been sent already
+  notifydb: ./notifications-{email}.db # State file used to remember which notifications has been sent already
 ```
 
 2) Run it
 ```
-nohup python3 luxmedSnip.py -c /path/to/my_favourite_surgeon.yml &
+nohup python3 luxmed_sniper.py -c /path/to/luxmed_john.yml &
 ```
 or you can split the configuration into separate users/doctors/providers config files
 ```
-nohup python3 luxmedSnip.py -c user_config.yml my_favourite_surgeon.yml &
+nohup python3 luxmed_sniper.py -c user_config.yml luxmed_john.yml &
 ```
 3) Wait for new appointment notifications in your pushover app on mobile :)!
