+++ luxmed_sniper.yaml
@@ -1,0 +1,40 @@
+luxmed:
+  email: EMAIL
+  password: PASSWORD
+luxmedsniper:
+  # (cityId, serviceVariantId, facilitiesIds, doctorsIds) -1 means any.
+  # You can get those ids by reading form data sent to https://portalpacjenta.luxmed.pl/PatientPortal/Reservations/Reservation/PartialSearch
+  # on https://portalpacjenta.luxmed.pl/PatientPortal/Reservations/Reservation/Search by Chrome dev tools
+  doctor_locators:
+    - id: 1*7409*-1*-1
+      name: Ala pediatra
+      enabled: False
+    - id: 1*7681*-1*-1
+      name: Ala okulista
+      enabled: True
+  lookup_time_days: 14 # How many days from now should script look at.
+  facilities_ids: []
+  notification_provider: # console, console_async, slack, pushover, pushbullet, gi, telegram
+    - telegram
+pushover:
+  user_key: # Your pushover.net user key
+  api_token:  # pushover.net App API Token
+  message_template: "{name} - {AppointmentDate} at {ClinicPublicName} - {DoctorName}"
+  title: "New Lux Med visit available!" # Pushover message topic
+slack:
+  api_token: # Your slack application token
+  channel: # notification channel
+  message_template: "{name} - {AppointmentDate} at {ClinicPublicName} - {DoctorName}"
+pushbullet:
+  access_token: # Your pushbullet access token
+  message_template: "{name} - {AppointmentDate} at {ClinicPublicName} - {DoctorName}"
+  title: "New Lux Med visit available!"
+telegram:
+  tele_conf_path: telegram-luxmed-notifier.conf # Your telegram_send config file path
+  message_template: "{name} - {AppointmentDate} at {ClinicPublicName} - {DoctorName}"
+console:
+  message_template: "{name} - {AppointmentDate} at {ClinicPublicName} - {DoctorName}"
+console_async:
+  message_template: "{name} - {AppointmentDate} at {ClinicPublicName} - {DoctorName}"
+misc:
+  notifydb_template: ./notifications-{email}.db # State file used to remember which notifications have been sent already
